[20221024] Member, Project, Requirement 엔티티&레포지토리 추가 및 Config 설정파일 추가

See merge request s07-final/S07P31B203!6

+++ back/myini-api/src/main/java/com/ssafy/myini/member/service/MemberService.java
@@ -1,5 +1,5 @@
 package com.ssafy.myini.member.service;
 
 public interface MemberService {
-    String generateToken(int userId);
+    String generateToken(Long userId);
 }
